Descripción de refactorizacion
