+++ app/src/main/java/com/jebereal/jeberealapp/presentation/screens/concertApp/PaymentScreen.kt
@@ -1,0 +1,81 @@
+package com.jebereal.jeberealapp.presentation.screens.concertApp
+
+import android.widget.Toast
+import androidx.compose.foundation.focusable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Button
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextField
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.unit.dp
+import com.jebereal.jeberealapp.util.showNotification
+
+
+@Composable
+fun PaymentScreen(totalAmount: Double, onPay: () -> Unit) {
+    var cardNumber by remember { mutableStateOf("") }
+    var cardHolderName by remember { mutableStateOf("") }
+
+    // Obtener el contexto actual
+    val context = LocalContext.current
+
+    Column(
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(16.dp),
+        verticalArrangement = Arrangement.Center
+    ) {
+        Text(
+            text = "Total: $$totalAmount",
+            style = MaterialTheme.typography.headlineMedium,
+            modifier = Modifier.padding(bottom = 16.dp)
+        )
+
+        TextField(
+            value = cardNumber,
+            onValueChange = { cardNumber = it },
+            label = { Text("Card Number") },
+            modifier = Modifier
+                .fillMaxWidth()
+                .focusable(true)
+        )
+
+        Spacer(modifier = Modifier.height(16.dp))
+
+        TextField(
+            value = cardHolderName,
+            onValueChange = { cardHolderName = it },
+            label = { Text("Cardholder Name") },
+            modifier = Modifier
+                .fillMaxWidth()
+                .focusable(true)
+        )
+
+        Spacer(modifier = Modifier.height(32.dp))
+
+        Button(
+            onClick = {
+                showNotification(context)
+                // Mostrar un toast con el monto total y el mensaje "Success"
+                Toast.makeText(context, "Success! Total: $$totalAmount", Toast.LENGTH_LONG).show()
+                onPay()
+            },
+            modifier = Modifier.fillMaxWidth()
+        ) {
+            Text("Pay Now")
+        }
+    }
+}
